refactor: FSD 아키텍처에 따른 settings 및 auth 관련 파일 구조 _FSD suffix에 맞게 경로 수정
