--- omechu-app/src/app/(routes)/(auth)/sign-up/page.tsx
@@ -1,152 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { FormProvider, useForm } from "react-hook-form";
-
-import BottomButton from "@/components/common/button/BottomButton";
-import ModalWrapper from "@/components/common/ModalWrapper";
-import Toast from "@/components/common/Toast";
-import { TERMS_CONFIG } from "@/components/settings/terms.config";
-import {
-  signupSchema,
-  type SignupFormValues,
-} from "@/entities/user/model/auth.schema";
-import { useSignupMutation, useLoginMutation } from "@/entities/user/lib/hooks/useAuth";
-import { useAuthStore } from "@/entities/user/model/auth.store";
-import { ApiClientError } from "@/entities/user/api/authApi";
-
-import SignUpForm from "@/widgets/auth/sign-up-form/ui/SignUpForm";
-import TermsModal from "@/widgets/auth/sign-up-form/ui/TermsModal";
-
-type ModalType = "service" | "privacy" | "location";
-
-/** 모달 타입과 약관 설정 매핑 */
-const MODAL_TO_TERMS_TYPE = {
-  service: "service",
-  privacy: "personal-info",
-  location: "location-info",
-} as const;
-
-/** 모달 타입과 폼 필드 매핑 */
-const MODAL_TO_FORM_FIELD = {
-  service: "termsService",
-  privacy: "termsPrivacy",
-  location: "termsLocation",
-} as const;
-
-export default function SignUpPage() {
-  const [activeModal, setActiveModal] = useState<ModalType | null>(null);
-  const [toastMessage, setToastMessage] = useState("");
-  const [showToast, setShowToast] = useState(false);
-  const router = useRouter();
-  const { setPassword } = useAuthStore();
-  const { mutate: signup, isPending: isSigningUp } = useSignupMutation();
-  const { mutateAsync: loginAsync } = useLoginMutation();
-
-  const methods = useForm<SignupFormValues>({
-    resolver: zodResolver(signupSchema),
-    mode: "onChange",
-    defaultValues: {
-      email: "",
-      password: "",
-      passwordConfirm: "",
-      verificationCode: "",
-      termsService: false,
-      termsPrivacy: false,
-      termsLocation: false,
-      termsAge: false,
-    },
-  });
-
-  const {
-    handleSubmit,
-    setValue,
-    formState: { isValid },
-  } = methods;
-
-  const triggerToast = (msg: string) => {
-    setToastMessage(msg);
-    setShowToast(true);
-    setTimeout(() => setShowToast(false), 2500);
-  };
-
-  const onSubmit = (data: SignupFormValues) => {
-    signup(data, {
-      onSuccess: async () => {
-        try {
-          // 회원가입 직후 자동 로그인하여 토큰을 확보 → 온보딩 완료 API에서 401 방지
-          await loginAsync({ email: data.email, password: data.password });
-          setPassword(data.password);
-          router.push("/onboarding/1");
-        } catch (e: any) {
-          // 자동 로그인 실패 시에도 온보딩으로 이동하되, 안내 토스트 노출
-          triggerToast(
-            e?.message ||
-              "자동 로그인에 실패했습니다. 로그인 후 계속 진행해 주세요.",
-          );
-          router.push("/onboarding/1");
-        }
-      },
-      onError: (error: unknown) => {
-        const e = error as ApiClientError & { code?: string };
-        const msg: string = e?.message || "회원가입에 실패했습니다.";
-        triggerToast(msg);
-      },
-    });
-  };
-
-  /** 약관 동의 확인 핸들러 */
-  const handleTermsConfirm = () => {
-    if (activeModal) {
-      const termKey = MODAL_TO_FORM_FIELD[activeModal];
-      setValue(termKey, true, { shouldValidate: true });
-    }
-    setActiveModal(null);
-  };
-
-  return (
-    <FormProvider {...methods}>
-      <div className="flex h-screen flex-col">
-        <header className="px-4 py-5 text-center">
-          <h1 className="py-10 text-xl font-bold text-grey-darker">
-            회원 정보를 입력해 주세요
-          </h1>
-        </header>
-
-        <main className="flex-1 overflow-y-auto px-5">
-          <SignUpForm
-            setActiveModal={setActiveModal}
-            onSubmit={handleSubmit(onSubmit)}
-          />
-        </main>
-
-        <footer className="w-full pb-[env(safe-area-inset-bottom)]">
-          <BottomButton
-            type="submit"
-            form="signup-form"
-            disabled={!isValid || isSigningUp}
-          >
-            {isSigningUp ? "가입하는 중..." : "가입하기"}
-          </BottomButton>
-        </footer>
-
-        {/* 약관 모달 */}
-        {activeModal && (
-          <ModalWrapper>
-            <TermsModal
-              title={TERMS_CONFIG[MODAL_TO_TERMS_TYPE[activeModal]].title}
-              terms={TERMS_CONFIG[MODAL_TO_TERMS_TYPE[activeModal]].data}
-              onConfirm={handleTermsConfirm}
-              onClose={() => setActiveModal(null)}
-            />
-          </ModalWrapper>
-        )}
-        
-        <Toast message={toastMessage} show={showToast} className="bottom-20" />
-      </div>
-    </FormProvider>
-  );
-}
