--- omechu-app/src/app/(routes)/settings/account-settings/change-password/page.tsx
@@ -1,5 +1,0 @@
-import ChangePasswordClient from "@/widgets/change-password/ui/ChangePasswordClient";
-
-export default function ChangePasswordPage() {
-  return <ChangePasswordClient />;
-}
