--- omechu-app/src/app/(routes)/settings/account-settings/page.tsx
@@ -1,5 +1,0 @@
-import AccountSettingsClient from "@/widgets/account-settings/ui/AccountSettingsClient";
-
-export default function AccountSettingsPage() {
-  return <AccountSettingsClient />;
-}
