--- omechu-app/src/app/(routes)/settings/page.tsx
@@ -1,5 +1,0 @@
-import SettingsClient from "@/widgets/settings-menu/ui/SettingsClient";
-
-export default function SettingsPage() {
-  return <SettingsClient />;
-}
