+++ omechu-app/src/app_FSD/(auth)/forgot-password/sent/components/EmailSentMessage.tsx
