+++ omechu-app/src/app_FSD/settings/account-settings/change-password/page.tsx
@@ -1,0 +1,5 @@
+import ChangePasswordClient from "@/widgets_FSD/change-password/ui/ChangePasswordClient";
+
+export default function ChangePasswordPage() {
+  return <ChangePasswordClient />;
+}
