+++ omechu-app/src/app_FSD/settings/account-settings/page.tsx
@@ -1,0 +1,5 @@
+import AccountSettingsClient from "@/widgets_FSD/account-settings/ui/AccountSettingsClient";
+
+export default function AccountSettingsPage() {
+  return <AccountSettingsClient />;
+}
