+++ omechu-app/src/app_FSD/settings/page.tsx
@@ -1,0 +1,5 @@
+import SettingsClient from "@/widgets_FSD/settings-menu/ui/SettingsClient";
+
+export default function SettingsPage() {
+  return <SettingsClient />;
+}
