+++ omechu-app/src/app_FSD/settings/terms/[type]/page.tsx
