+++ omechu-app/src/app_FSD/settings/terms/page.tsx
