+++ omechu-app/src/widgets_FSD/auth/sign-up-form/ui/TermsModal.tsx
